[Add] Debugger in Engine and Editor
Major:
The debugger in the engine side redirects the log messages to a listener that also derives from the Logger class.
Therefore, on the side of the Editor, we create a class EditorLog that derives from Logger and make it the listener of the EngineLog. Then, EditorLog forwards the message to the Log widget.
Important to link the EditorLog with the LogWidget; and the EditorLog with the EngineLog.

Minor:
Commented out the widgets that were relevant in the previous version since they will eventually disappear. I kept the files for now to have the content of the code for reference.