[Retro] Going back to libxml2
There seems to be an issue with expat library returning a null pointer when loading the model in a sbml file.
No solution were found so we go back to libxml2 for now untill we find a fix.
For a reminder, we need to use expat if we want the CI to work because libxml2 fails for MacOs and Linux.