[Fix] SBML Model
This is not related to the code but I make it a commit to keep trace of the issue.
The problem was that sode kinetic laws in the model were using nodes "Times" with only a single child. This is like writing "x = y * "
I suppose it could be implicitely converted to something like "x = y * 1" but there is no reason to put that burden on the parser with additional checks because the libSbml doc defines a product as a binary operation with two elements (https://sbml.org/software/libsbml/5.18.0/docs/formatted/cpp-api/class_a_s_t_node.html). So, instead, I corrected by hand the structure of the model.