[Fix] Remove useless resizing of vectors on sbml import
On import, we were reserving the space for species, parameters and reactions list before parsing the corresponding data in the SBML File.
This was legacy code and useless. Moreover, we were doing it wrongly bu using .resize() instead of .reserve() which lead to having twice as much as items as we should have.